Move array.rs around
